# src/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API de gestion para restaurante 🍽️🍽️",
    "description": "Documentacion generada automaticamente por swagger-autogen",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/repartidores/": {
      "get": {
        "tags": [
          "Repartidores"
        ],
        "summary": "Lista todos los repartidores",
        "description": "Permite filtrar por disponibilidad con ?activo=true",
        "parameters": [
          {
            "name": "activo",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de repartidores obtenida correctamente"
          }
        }
      },
      "post": {
        "tags": [
          "Repartidores"
        ],
        "summary": "Registra un nuevo repartidor",
        "description": "Requiere nombre, vehiculo, telefono y activo en el body",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nombre": {
                  "example": "any"
                },
                "vehiculo": {
                  "example": "any"
                },
                "telefono": {
                  "example": "any"
                },
                "activo": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Repartidor creado correctamente"
          },
          "400": {
            "description": "Datos inválidos o incompletos"
          }
        }
      }
    },
    "/repartidores/{id}": {
      "get": {
        "tags": [
          "Repartidores"
        ],
        "summary": "Obtiene un repartidor por ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID numérico del repartidor"
          }
        ],
        "responses": {
          "200": {
            "description": "Repartidor encontrado"
          },
          "404": {
            "description": "Repartidor no encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Repartidores"
        ],
        "summary": "Actualiza datos, vehículo o disponibilidad de un repartidor",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID numérico del repartidor"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nombre": {
                  "example": "any"
                },
                "vehiculo": {
                  "example": "any"
                },
                "telefono": {
                  "example": "any"
                },
                "activo": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Repartidor actualizado correctamente"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Repartidor no encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "Repartidores"
        ],
        "summary": "Da de baja a un repartidor",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID numérico del repartidor"
          }
        ],
        "responses": {
          "204": {
            "description": "Repartidor eliminado correctamente"
          },
          "404": {
            "description": "Repartidor no encontrado"
          }
        }
      }
    },
    "/productos/": {
      "get": {
        "tags": [
          "Productos"
        ],
        "description": "Obtiene la lista de productos, con filtros opcionales por nombre, categoria y disponible",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Productos"
        ],
        "description": "Crea un nuevo producto, validando que el precio sea mayor a 0",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nombre": {
                  "example": "any"
                },
                "precio": {
                  "example": "any"
                },
                "categoria": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/productos/{id}": {
      "get": {
        "tags": [
          "Productos"
        ],
        "description": "Obtiene un producto especifico por su id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "Productos"
        ],
        "description": "Modifica el precio, nombre, categoria o disponibilidad de un producto existente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Productos"
        ],
        "description": "Elimina un producto del catalogo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    }
  }
}